Added the vendor form folder and index.js file
I used reactstrap to have prestyled components, and used a Select HTML5 element for a drop down so the couple can choose which vendor group the vendor belongs to. I created two helper functions handleChange and handleSubmit. The handleChange function should allow the user to select an option from a dropdown, and the handleSubmit function is still needing to be worked on, which I will do in a future commit.